add: EqueipmentFile
