+++ Spartale/Source/GameLogic/Items/EquipItem.cpp
