+++ Spartale/Source/GameLogic/Items/EquipItem.h
