+++ Spartale/Source/GameLogic/Items/EquipSlotType.h
