+++ Spartale/Source/GameLogic/Items/EquipmentSlot.h
