+++ Spartale/Source/GameLogic/Items/ItemBase.h
