+++ Spartale/Spartale.vcxproj
@@ -128,41 +128,55 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="AbilitySystemComponent.cpp" />
-    <ClCompile Include="AB_NormalAttack.cpp" />
-    <ClCompile Include="Actor.cpp" />
-    <ClCompile Include="AttributeSet.cpp" />
-    <ClCompile Include="BattleManager.cpp" />
-    <ClCompile Include="ConsoleUtils.cpp" />
-    <ClCompile Include="DamageUtils.cpp" />
-    <ClCompile Include="EquipItem.cpp" />
-    <ClCompile Include="EquipmentSlot.cpp" />
-    <ClCompile Include="GameplayAbility.cpp" />
-    <ClCompile Include="GameplayEffect.cpp" />
-    <ClCompile Include="ItemBase.cpp" />
-    <ClCompile Include="Monster.cpp" />
-    <ClCompile Include="Player.cpp" />
+    <ClCompile Include="Source\Core\Actor.cpp" />
+    <ClCompile Include="Source\Core\GameWorld.cpp" />
+    <ClCompile Include="Source\Framework\AbilitySystem\AbilitySystemComponent.cpp" />
+    <ClCompile Include="Source\Framework\AbilitySystem\AttributeSet.cpp" />
+    <ClCompile Include="Source\Framework\AbilitySystem\GameplayAbility.cpp" />
+    <ClCompile Include="Source\Framework\AbilitySystem\GameplayEffect.cpp" />
+    <ClCompile Include="Source\GameLogic\BattleManager.cpp" />
+    <ClCompile Include="Source\GameLogic\Items\EquipItem.cpp" />
+    <ClCompile Include="Source\GameLogic\Items\EquipmentSlot.cpp" />
+    <ClCompile Include="Source\GameLogic\Items\HealthPotion.cpp" />
+    <ClCompile Include="Source\GameLogic\Items\ItemBase.cpp" />
+    <ClCompile Include="Source\GameLogic\Map\Field.cpp" />
+    <ClCompile Include="Source\GameLogic\Skills\AB_NormalAttack.cpp" />
+    <ClCompile Include="Source\GameLogic\Skills\AB_PoisonCloud.cpp" />
+    <ClCompile Include="Source\GameLogic\Units\Monster.cpp" />
+    <ClCompile Include="Source\GameLogic\Units\Player.cpp" />
+    <ClCompile Include="Source\Utils\ConsoleRenderer.cpp" />
+    <ClCompile Include="Source\Utils\ConsoleUtils.cpp" />
+    <ClCompile Include="Source\Utils\DamageUtils.cpp" />
     <ClCompile Include="Spartale.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="AbilitySystemComponent.h" />
-    <ClInclude Include="AB_NormalAttack.h" />
-    <ClInclude Include="Actor.h" />
-    <ClInclude Include="AttributeSet.h" />
-    <ClInclude Include="BattleManager.h" />
-    <ClInclude Include="ConsoleUtils.h" />
-    <ClInclude Include="DamageUtils.h" />
-    <ClInclude Include="EquipItem.h" />
-    <ClInclude Include="EquipmentSlot.h" />
-    <ClInclude Include="EquipSlotType.h" />
-    <ClInclude Include="GameplayAbility.h" />
-    <ClInclude Include="GameplayEffect.h" />
-    <ClInclude Include="Global.h" />
-    <ClInclude Include="HealthPotion.h" />
-    <ClInclude Include="ItemBase.h" />
-    <ClInclude Include="Monster.h" />
-    <ClInclude Include="Player.h" />
-    <ClInclude Include="Sword.h" />
+    <ClInclude Include="include\json.hpp" />
+    <ClInclude Include="Source\Core\Actor.h" />
+    <ClInclude Include="Source\Core\GameWorld.h" />
+    <ClInclude Include="Source\Core\Global.h" />
+    <ClInclude Include="Source\Framework\AbilitySystem\AbilitySystemComponent.h" />
+    <ClInclude Include="Source\Framework\AbilitySystem\AttributeSet.h" />
+    <ClInclude Include="Source\Framework\AbilitySystem\GameplayAbility.h" />
+    <ClInclude Include="Source\Framework\AbilitySystem\GameplayEffect.h" />
+    <ClInclude Include="Source\GameLogic\BattleManager.h" />
+    <ClInclude Include="Source\GameLogic\Items\EquipItem.h" />
+    <ClInclude Include="Source\GameLogic\Items\EquipmentSlot.h" />
+    <ClInclude Include="Source\GameLogic\Items\EquipSlotType.h" />
+    <ClInclude Include="Source\GameLogic\Items\HealthPotion.h" />
+    <ClInclude Include="Source\GameLogic\Items\ItemBase.h" />
+    <ClInclude Include="Source\GameLogic\Items\Sword.h" />
+    <ClInclude Include="Source\GameLogic\Map\Field.h" />
+    <ClInclude Include="Source\GameLogic\Skills\AB_NormalAttack.h" />
+    <ClInclude Include="Source\GameLogic\Skills\AB_PoisonCloud.h" />
+    <ClInclude Include="Source\GameLogic\Units\Monster.h" />
+    <ClInclude Include="Source\GameLogic\Units\Player.h" />
+    <ClInclude Include="Source\Utils\ConsoleRenderer.h" />
+    <ClInclude Include="Source\Utils\ConsoleUtils.h" />
+    <ClInclude Include="Source\Utils\DamageUtils.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Data\Field_1.json" />
+    <None Include="Data\Town.json" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
